% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for the stan model}
\usage{
prepare_data(
  linelist,
  pairs,
  si = list(mean_mean = 9.2, mean_sd = 1, sd_mean = 4.4, sd_sd = 1),
  od = list(mu_mean = log(8.5), mu_sd = 1, sigma_mean = log(2.5), sigma_sd = 1),
  ip = list(mu_mean = log(7), mu_sd = 1, sigma_mean = log(2.5), sigma_sd = 1),
  init = 21
)
}
\arguments{
\item{linelist}{a linelist data frame}

\item{pairs}{a pairs data frame}

\item{od}{onset-to-death parameters}

\item{ip}{incubation period parameters}

\item{gi}{generation interval parameters}

\item{max_ip}{maximum possible incubation period}
}
\value{
a list ready for stan
}
\description{
Prepare data for the stan model
}
\author{
Sebastian Funk
}
