% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_init.r
\name{prepare_init}
\alias{prepare_init}
\title{Prepare initial conditions for the stan model}
\usage{
prepare_init(data)
}
\arguments{
\item{data}{a list of stan data}
}
\value{
a function ready for stan
}
\description{
Prepare initial conditions for the stan model
}
\author{
Sebastian Funk
}
